requirements.txt
streamlit
python-dotenv
huggingface_hub
plotly
yfinance